refactor: move the feature check inside the hooks
